Bump project version to 0.0.2 and update MyECS dependency to version 0.13.1; adjust formatting in AutoRefl files for consistency

+++ src/App/Editor/Systems/ProjectViewerSystem.cpp
@@ -15,7 +15,8 @@ namespace My::MyGE::detail {
 bool IsFolderLeaf(const xg::Guid& folderGuid) {
   const auto& tree = AssetMngr::Instance().GetAssetTree();
   auto target = tree.find(folderGuid);
-  if (target == tree.end()) return true;
+  if (target == tree.end())
+    return true;
 
   for (const auto& child : target->second) {
     if (std::filesystem::is_directory(
@@ -40,7 +41,8 @@ bool IsParentFolder(const std::filesystem::path& x,
 }
 
 void NameShrink(std::string& name, size_t maxlength) {
-  if (name.size() <= maxlength) return;
+  if (name.size() <= maxlength)
+    return;
   name.resize(maxlength);
   name[maxlength - 2] = '.';
   name[maxlength - 1] = '.';
@@ -60,7 +62,8 @@ void ProjectViewerSystemPrintChildren(ProjectViewer* viewer,
 
   for (const auto& child : children) {
     const auto& path = AssetMngr::Instance().GUIDToAssetPath(child);
-    if (!std::filesystem::is_directory(path)) continue;
+    if (!std::filesystem::is_directory(path))
+      continue;
     bool isFolderLeaf = IsFolderLeaf(child);
 
     auto name = path.filename();
@@ -77,7 +80,8 @@ void ProjectViewerSystemPrintChildren(ProjectViewer* viewer,
     bool nodeOpen = ImGui::TreeNodeEx(AssetMngr::Instance().LoadAsset(path),
                                       nodeFlags, "%s", name.string().c_str());
 
-    if (ImGui::IsItemClicked()) viewer->selectedFolder = child;
+    if (ImGui::IsItemClicked())
+      viewer->selectedFolder = child;
 
     if (nodeOpen && !isFolderLeaf) {
       ProjectViewerSystemPrintChildren(viewer, child);
@@ -89,7 +93,8 @@ void ProjectViewerSystemPrintChildren(ProjectViewer* viewer,
 void ProjectViewerSystemPrintFolder(Inspector* inspector,
                                     ProjectViewer* viewer) {
   const auto& tree = AssetMngr::Instance().GetAssetTree();
-  if (!viewer->selectedFolder.isValid()) return;
+  if (!viewer->selectedFolder.isValid())
+    return;
 
   std::vector<std::filesystem::path> paths;
   auto curPath = AssetMngr::Instance().GUIDToAssetPath(viewer->selectedFolder);
@@ -148,11 +153,12 @@ void ProjectViewerSystemPrintFolder(Inspector* inspector,
   ImVec2 button_sz(64, 64);
   float window_visible_x2 =
       ImGui::GetWindowPos().x + ImGui::GetWindowContentRegionMax().x;
-  size_t idx = 0;
+  UINT idx = 0;
   // pass 1 : folder
   for (const auto& child : children) {
     const auto& path = AssetMngr::Instance().GUIDToAssetPath(child);
-    if (!std::filesystem::is_directory(path)) continue;
+    if (!std::filesystem::is_directory(path))
+      continue;
     ImGui::PushID(idx);
     auto name = path.filename();
 
@@ -197,7 +203,8 @@ void ProjectViewerSystemPrintFolder(Inspector* inspector,
   // pass 2 : files
   for (const auto& child : children) {
     const auto& path = AssetMngr::Instance().GUIDToAssetPath(child);
-    if (std::filesystem::is_directory(path)) continue;
+    if (std::filesystem::is_directory(path))
+      continue;
     ImGui::PushID(idx);
     auto name = path.stem();
     auto ext = path.extension();
@@ -269,7 +276,8 @@ void ProjectViewerSystem::OnUpdate(MyECS::Schedule& schedule) {
     auto viewer = w->entityMngr.GetSingleton<ProjectViewer>();
     auto inspector = w->entityMngr.GetSingleton<Inspector>();
 
-    if (!viewer) return;
+    if (!viewer)
+      return;
 
     if (ImGui::Begin("Project"))
       detail::ProjectViewerSystemPrintChildren(viewer, xg::Guid{});
